+++ apps/property-manager/app/modules/tenants/apply/steps/step2.tsx
@@ -1,11 +1,7 @@
-import { zodResolver } from '@hookform/resolvers/zod'
 import { REGEXP_ONLY_DIGITS } from 'input-otp'
 import { ArrowLeft, ArrowRight, Loader2 } from 'lucide-react'
 import { useEffect, useState } from 'react'
-import { useForm } from 'react-hook-form'
-import { useNavigate } from 'react-router'
-import { z } from 'zod'
-import { useTenantApplicationContext } from './context'
+import { useTenantApplicationContext } from '../context'
 import { Button } from '~/components/ui/button'
 import {
 	InputOTP,
@@ -18,38 +14,80 @@ import {
 	TypographySmall,
 } from '~/components/ui/typography'
 
-const ValidationSchema = z.object({
-	phone: z
-		.string({ error: 'Phone number is required' })
-		.min(9, 'Please enter a valid phone number'),
-	otp: z
-		.string({ error: 'OTP is required' })
-		.length(4, 'Please enter a valid 4-digit OTP'),
-})
+const tenantApplicationData: TenantApplication = {
+	id: 'ta_001',
+	on_boarding_method: 'SELF',
 
-type FormSchema = z.infer<typeof ValidationSchema>
+	first_name: 'Kwame',
+	other_names: 'Nana',
+	last_name: 'Mensah',
+	email: 'kwame.mensah@example.com',
+	phone: '+233501234567',
+	gender: 'MALE',
+	date_of_birth: '1994-06-15',
+	nationality: 'Ghanaian',
+	marital_status: 'SINGLE',
+
+	profile_photo_url: null,
+
+	id_type: 'NATIONAL_ID',
+	id_number: 'GHA-123456789',
+	id_front_url: null,
+	id_back_url: null,
+
+	status: 'TenantApplication.Status.InProgress',
+
+	current_address: 'East Legon, Accra',
+	emergency_contact_name: 'Ama Mensah',
+	emergency_contact_phone: '+233241112223',
+	relationship_to_emergency_contact: 'Sister',
+
+	employment_type: 'WORKER',
+	occupation: 'Software Developer',
+	employer: 'Tech Solutions Ltd',
+	occupation_address: 'Airport, Accra',
+	proof_of_income_url: null,
+
+	created_by: null,
+	created_by_id: 'user_001',
+
+	completed_at: null,
+	completed_by_id: null,
+	completed_by: null,
+
+	cancelled_at: null,
+	cancelled_by_id: null,
+	cancelled_by: null,
+
+	desired_unit_id: 'unit_123',
+	desired_unit: {
+		id: 'unit_123',
+		name: '2 Bedroom Apartment',
+		// add other required Unit fields here
+	} as PropertyUnit,
+
+	previous_landlord_name: 'Mr. Boateng',
+	previous_landlord_phone: '+233209998887',
+	previous_tenancy_period: 'Jan 2021 - Dec 2023',
+
+	created_at: new Date('2024-01-10T10:00:00Z'),
+	updated_at: new Date('2024-01-15T14:30:00Z'),
+}
 
 const isSubmitting = false
 const error = null
 
-export function TenantApplicationOTPValidationModule({
-	referredBy,
-	unitId,
-}: {
-	referredBy: string | null
-	unitId: string | null
-}) {
-	const navigate = useNavigate()
+export function Step2() {
 	const [otp, setOtp] = useState('')
 	const [canResend, setCanResend] = useState(true)
 	const [resendCountdown, setResendCountdown] = useState(0)
 
-	const { goBack, goNext, formData, unblockNavigation } =
+	const { goBack, goNext, goToPage, formData, updateFormData, allowEdit } =
 		useTenantApplicationContext()
 
-	const rhfMethods = useForm<FormSchema>({
-		resolver: zodResolver(ValidationSchema),
-	})
+	// const rhfMethods = useForm<FormSchema>({
+	// 	resolver: zodResolver(ValidationSchema),
+	// })
 
 	useEffect(() => {
 		let t: NodeJS.Timeout | null = null
@@ -72,23 +110,65 @@ export function TenantApplicationOTPValidationModule({
 
 	useEffect(() => {
 		if (otp.length === 4) {
-			verify()
+			void onSubmit()
 		}
 	}, [otp])
 
 	const verify = () => {
 		//    ToDO: verify OTP logic (Mutate to backend)
+	}
+
+	const phoneLookUp = () => {
+		//    ToDO: Phone lookup logic to get user data
 		if (otp !== '1234') {
-			// Clear form data and set bypass flag to allow navigation without prompting
-			unblockNavigation()
-			// Small delay to ensure state updates before navigation
-			setTimeout(() => {
-				void navigate(
-					`/tenants/apply/new?unit=${unitId}&referred_by=${referredBy}`,
-				)
-			}, 0)
+			return null
 		} else {
+			return tenantApplicationData
+		}
+	}
+
+	const onSubmit = async () => {
+		verify()
+		const lookUpData = phoneLookUp()
+		if (!lookUpData) {
+			allowEdit(true)
 			goNext()
+		} else {
+			updateFormData({
+				// step3 data
+				first_name: lookUpData.first_name,
+				other_names: lookUpData.other_names,
+				last_name: lookUpData.last_name,
+				email: lookUpData.email,
+				phone: lookUpData.phone,
+				current_address: lookUpData.current_address,
+				profile_photo_url: lookUpData.profile_photo_url,
+				date_of_birth: lookUpData.date_of_birth?.toString(),
+				gender: lookUpData.gender,
+				marital_status: lookUpData.marital_status,
+
+				// step4 data
+				nationality: lookUpData.nationality,
+				id_type: lookUpData.id_type,
+				id_number: lookUpData.id_number,
+				id_front_url: lookUpData.id_front_url,
+				id_back_url: lookUpData.id_back_url,
+
+				// step5 data
+				emergency_contact_name: lookUpData.emergency_contact_name,
+				relationship_to_emergency_contact:
+					lookUpData.relationship_to_emergency_contact,
+				emergency_contact_phone: lookUpData.emergency_contact_phone,
+				employment_type: lookUpData.employment_type,
+				occupation:
+					lookUpData.employment_type === 'STUDENT'
+						? lookUpData.employment_type
+						: lookUpData.occupation,
+				employer: lookUpData.employer,
+				occupation_address: lookUpData.occupation_address,
+				proof_of_income_url: lookUpData.proof_of_income_url,
+			})
+			goToPage(6)
 		}
 	}
 
@@ -159,7 +239,7 @@ export function TenantApplicationOTPValidationModule({
 					<Button
 						size="lg"
 						variant="default"
-						onClick={verify}
+						onClick={onSubmit}
 						disabled={otp.length < 4 || isSubmitting}
 						className="w-full bg-rose-600 hover:bg-rose-700 md:w-auto"
 					>
