+++ apps/property-manager/app/modules/tenants/apply/steps/step4.tsx
@@ -58,7 +58,7 @@ const idTypes: Array<{
 
 export type FormSchema = z.infer<typeof ValidationSchema>
 
-export function Step2() {
+export function Step4() {
 	const { goBack, goNext, formData, updateFormData } =
 		useTenantApplicationContext()
