+++ apps/property-manager/app/routes/tenants.apply._index.tsx
@@ -1,6 +1,7 @@
 import { redirect } from 'react-router'
 import type { Route } from './+types/apply._index'
 import { createTenantApplication } from '~/api/tenant-applications'
+import { getPropertyUnitForServer } from '~/api/units/server'
 import { environmentVariables } from '~/lib/actions/env.server'
 import { replaceNullUndefinedWithUndefined } from '~/lib/actions/utils.server'
 import { APP_NAME } from '~/lib/constants'
@@ -9,8 +10,24 @@ import { getSocialMetas } from '~/lib/seo'
 import { TenantApplyModule } from '~/modules'
 
 export async function loader({ request }: Route.LoaderArgs) {
+	const baseUrl = environmentVariables().API_ADDRESS
+	const url = new URL(request.url)
+	const referredBy = url.searchParams.get('referred_by')
+	const unitId = url.searchParams.get('unit')
+
+	const unit = await getPropertyUnitForServer(
+		{ unit_id: unitId as string },
+		{
+			baseUrl,
+		},
+	)
+
 	return {
 		origin: getDomainUrl(request),
+		referredBy,
+		unitId,
+		unit,
+		isValidUrl: !!(referredBy && unitId),
 	}
 }
