--- apps/property-manager/app/routes/tenants.apply.new.tsx
@@ -1,26 +1,0 @@
-import type { Route } from './+types/apply._index'
-import { APP_NAME } from '~/lib/constants'
-import { getDisplayUrl, getDomainUrl } from '~/lib/misc'
-import { getSocialMetas } from '~/lib/seo'
-import { NewTenantApplyModule } from '~/modules'
-
-export async function loader({ request }: Route.LoaderArgs) {
-	return {
-		origin: getDomainUrl(request),
-	}
-}
-
-export function meta({ loaderData, location }: Route.MetaArgs) {
-	const meta = getSocialMetas({
-		title: `Complete Tenant Application Details - ${APP_NAME}`,
-		url: getDisplayUrl({
-			origin: loaderData.origin,
-			path: location.pathname,
-		}),
-		origin: loaderData.origin,
-	})
-
-	return meta
-}
-
-export default NewTenantApplyModule
